Simplify taxon metadata wiring and segregate CLI options by workflow
Remove the TaxonProvider/TaxonSource/TaxonIdMetadataSource abstraction.
No production code ever registered a TaxonSource, so every lookup returned
empty and the chain-of-responsibility was dead weight. Drop the deferred
materialization in MasterMetadataProvider that existed solely to resolve
TaxonProvider from the validation context.

EmblEntryMetadataSource no longer takes a TaxonProvider; the db_xref
fallback now just recovers the raw taxon ID (matching the previous
effective behavior).

Segregate the CLI options per workflow:
- --master-entry moves to FileConversionCommand (GFF3<->FF conversion)
- --taxon-id moves to ValidationCommand (validation workflow)
