+++ src/main/java/uk/ac/ebi/embl/gff3tools/metadata/MasterMetadataProvider.java
@@ -13,7 +13,6 @@ package uk.ac.ebi.embl.gff3tools.metadata;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import uk.ac.ebi.embl.api.entry.Entry;
 import uk.ac.ebi.embl.gff3tools.validation.ContextProvider;
 import uk.ac.ebi.embl.gff3tools.validation.ValidationContext;
 
@@ -28,51 +27,11 @@ import uk.ac.ebi.embl.gff3tools.validation.ValidationContext;
  */
 public class MasterMetadataProvider implements ContextProvider<MasterMetadataProvider> {
 
-    private final List<MasterMetadataSource> sources;
-    private Entry pendingEmblMasterEntry;
-    private Long pendingTaxonId;
-
-    public MasterMetadataProvider() {
-        this.sources = new ArrayList<>();
-    }
+    private final List<MasterMetadataSource> sources = new ArrayList<>();
 
     @Override
     public MasterMetadataProvider get(ValidationContext context) {
-        if (pendingEmblMasterEntry != null) {
-            // Materialize the EMBL master entry source now that the context exists, so we
-            // can resolve a TaxonProvider from it instead of constructing one inline.
-            TaxonProvider taxonProvider = context.get(TaxonProvider.class);
-            sources.add(new EmblEntryMetadataSource(pendingEmblMasterEntry, taxonProvider));
-            pendingEmblMasterEntry = null;
-        }
-        if (pendingTaxonId != null) {
-            TaxonProvider taxonProvider = context.get(TaxonProvider.class);
-            sources.add(new TaxonIdMetadataSource(pendingTaxonId, taxonProvider));
-            pendingTaxonId = null;
-        }
         return this;
-    }
-
-    /**
-     * Defers building an {@link EmblEntryMetadataSource} until the validation context exists,
-     * so the source can use the registered {@link TaxonProvider}. The source is added on
-     * the first call to {@link #get(ValidationContext)}.
-     *
-     * <p>Use this for the EMBL flatfile master entry path. For pre-built sources (e.g. JSON),
-     * use {@link #addSource(MasterMetadataSource)} directly.
-     */
-    public void setEmblMasterEntry(Entry entry) {
-        this.pendingEmblMasterEntry = entry;
-    }
-
-    /**
-     * Defers building a {@link TaxonIdMetadataSource} until the validation context exists,
-     * so the source can resolve scientific name / lineage from the registered
-     * {@link TaxonProvider}. The source is added on the first call to
-     * {@link #get(ValidationContext)}.
-     */
-    public void setTaxonId(Long taxonId) {
-        this.pendingTaxonId = taxonId;
     }
 
     @Override
